pequeños cambios

+++ E07/E07.js
@@ -1,8 +1,7 @@
 function calcularPotencia() {
-    let base = parseFloat(document.getElementById('base').value);
-    let exponente = parseFloat(document.getElementById('exponente').value);
-    
-    
-    let resultado = Math.pow(base, exponente);
-    document.getElementById('resultado').innerText = `Resultado: ${resultado}`;
+  let base = parseInt(document.getElementById("base").value);
+  let exponente = parseInt(document.getElementById("exponente").value);
+
+  let resultado = Math.pow(base, exponente);
+  document.getElementById("resultado").innerText = `Resultado: ${resultado}`;
 }
